Added register and login functionalities, and separated responsibilities.
Updated name of folder to a2_group11 from projectfile. Edited user.html and login.html (register and login forms). Added users.py to separate responsibilities for user related blue prints such as display booking history, and create or update events.
