+++ a2_starter_code/a2_group11/main.py
@@ -1,0 +1,5 @@
+from website import create_app
+
+if __name__ == '__main__':
+    app = create_app()
+    app.run()
