+++ a2_starter_code/a2_group11/website/auth.py
@@ -1,0 +1,42 @@
+from flask import Blueprint, flash, render_template, request, url_for, redirect
+from flask_bcrypt import generate_password_hash, check_password_hash
+from flask_login import login_user, login_required, logout_user
+from .models import User
+from .forms import LoginForm, RegisterForm
+from . import db
+
+# Create a blueprint - make sure all BPs have unique names
+auth_bp = Blueprint('auth', __name__)
+
+# this is a hint for a login function
+@auth_bp.route('/login', methods=['GET', 'POST'])
+# view function
+def login():
+    login_form = LoginForm()
+    error = None
+    if login_form.validate_on_submit():
+        user_name = login_form.user_name.data
+        password = login_form.password.data
+        user = db.session.scalar(db.select(User).where(User.name==user_name))
+        if user is None:
+            error = 'Incorrect user name'
+        elif not check_password_hash(user.password_hash, password): # takes the hash and cleartext password
+            error = 'Incorrect password'
+        if error is None:
+            login_user(user)
+            nextp = request.args.get('next') # this gives the url from where the login page was accessed
+            print(nextp)
+            if next is None or not nextp.startswith('/'):
+                return redirect(url_for('index'))
+            return redirect(nextp)
+        else:
+            flash(error)
+    return render_template('user.html', form=login_form, heading='Login')
+
+@auth_bp.route('/register', methods=['GET','POST'])
+def register():
+    form = RegisterForm()
+    if form.validate_on_submit():
+        print('Successfully registered')
+        return redirect(url_for('auth.login'))
+    return render_template('user.html', form=form)
