+++ a2_starter_code/a2_group11/website/static/img/Friends.jpg
