+++ a2_starter_code/a2_group11/website/templates/events/update.html
@@ -1,0 +1,25 @@
+{% extends 'base.html' %}
+
+{% from 'bootstrap5/form.html' import render_form %}
+
+{% block content %}
+
+<!-- Container -->
+<div class="container-create">
+<!-- Event Creation and Update page Jumbotron -->
+<div class="p-5 creationjumbotronimage rounded-3">
+    <div class="jumbotronback container-fluid py-5 ">
+        <h1 class="display-5 fw-bold">Update an Event</h1>
+        <p class="col-md-8 fs-4">At FoodieVent, it's not just an event. Share memories, bring people
+            together. Enjoy food, drinks, and company. Update your pre-existing event here.</p>
+    </div>
+</div>
+<!-- Form -->
+    <div class="row mt-5">
+      <div class="col-12 col-md-10 col-lg-8 offset-md-2">
+        {{ render_form(form, button_map={'submit': 'success'}) }}
+      </div>
+    </div>
+  </div>
+
+{% endblock %}
