+++ a2_starter_code/a2_group11/website/templates/user.html
@@ -1,0 +1,24 @@
+{% extends "base.html" %}
+<!-- Title -->
+{% block title %}FoodieVent Register Page{% endblock %}
+{% from "bootstrap5/form.html" import render_form %}
+{% block content %}
+
+<h2 class="category-header">Register an account</h2>
+<div class="page-header">
+    <h4>{{heading}}</h4>
+</div>
+
+<!-- Form rendering is simple and easy if using WTForms and Bootstrap -->
+<div class="col-12 col-md-10 col-lg-8 offset-md-2">
+    {{render_form(form)}}
+</div>
+
+<!-- When you have Login form, then show the user a link to register -->
+{% if heading == 'Login' %}
+<div class="col-md-4">
+    <p>New user? <a href="{{url_for('register')}}">Click here</a> to register.</p>
+</div>
+{% endif %}
+
+{% endblock %}
