+++ a2_starter_code/a2_group11/website/views.py
@@ -1,0 +1,31 @@
+from flask import Blueprint, render_template, request, session
+from . models import Event
+from . import db
+from flask_login import login_required, current_user
+
+main_bp = Blueprint('main', __name__)
+
+@main_bp.route('/')
+def index():
+    events = db.session.scalars(db.select(Event)).all()    
+    return render_template('index.html', events=events)
+
+@main_bp.route('/display_event_details')
+def display_event_details():
+    return render_template('eventdetails.html')
+
+@main_bp.route('/login', methods = ['GET', 'POST'])
+def login():
+    email = request.values.get("email")
+    passwd = request.values.get("pwd")
+    print (f"Email: {email}\nPassword: {passwd}")
+    # store email in session
+    session['email'] = request.values.get('email')
+    return render_template('login.html')
+
+@main_bp.route('/logout')
+def logout():
+    if 'email' in session:
+        session.pop('email')
+    return 'User logged out'
+
