--- a2_starter_code/projectfile/main.py
@@ -1,5 +1,0 @@
-from website import create_app
-
-if __name__ == '__main__':
-    app = create_app()
-    app.run()
