--- a2_starter_code/projectfile/website/templates/user.html
@@ -1,21 +1,0 @@
-{% include "base.html" %}
-{% from "bootstrap5/form.html" import render_form %}
-{% block body %}
-
-<div class="page-header">
-    <h4>{{heading}}</h4>
-</div>
-
-<!-- Form rendering is simple and easy if using WTForms and Bootstrap -->
-<div class="col-md-6">
-    {{render_form(form)}}
-</div>
-
-<!-- When you have Login form, then show the user a link to register -->
-{% if heading == 'Login' %}
-<div class="col-md-4">
-    <p>New user? <a href="{{url_for('register')}}">Click here</a> to register.</p>
-</div>
-{% endif %}
-
-{% endblock %}
